Rename models to response
